refactor(discovery): rename arkham-discovery → mento-address-discovery and drop Arkham-only chain gate
The helper has always been provider-agnostic; only its name and a chain
gate were Arkham-flavored. Both consumer crons (Arkham enrichment +
MiniPay tagging) already pass `CELO_CHAIN_ID` for their own
provider-specific reasons, so the gate-and-throw inside the helper was
dead code. Move chain selection to the callers, rename file/test/Sentry
tag to match what the code actually does.
